backend/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Babel==4.0.0
Flask-CORS==5.0.1
Flask-Caching==2.3.0
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.1
requests==2.32.3
Werkzeug==3.1.3
pydantic[email]==2.7.1
PyJWT==2.8.0
reportlab==4.1.0
Faker==24.0.0
